Extend Jinja2 with {% markdown %} ... {% endmarkdown %}
This adds a {% markdown %} ... {% endmarkdown %} control structure to
Jinja2 that renders ... using Markdown.

The specific flavor of Markdown used is GitHub Flavored Markdown, with
all extensions except "tagfilter" and "tasklist" enabled, and with the
unsafe, smart, normalize, footnotes, and table-prefer-style-attributes
options.

To accommodate this change, the "lead" class is applied to any <p>
element within <main> that immediately follows an <h1>. And a Javascript
snippet is included in the base template that automatically adds
target="_blank" to all external links.

The About, Contact, Registration, Sponsorship Opportunities, and Venue
pages have been rewritten to use this feature.